# specs/openapi/gits-kno-api.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GITS Knowledge Engineering API",
    "version": "1.1.0",
    "description": "GITS 知识工程平台 API — V1.1"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/engagement/customer/{customerId}/operating-view": {
      "get": {
        "operationId": "getOperatingView",
        "tags": [
          "Customer Context"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/customerId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOperatingView"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/engagement/customer": {
      "post": {
        "operationId": "createCustomer",
        "tags": [
          "Customer Context"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerCreatedResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "get": {
        "operationId": "listCustomers",
        "tags": [
          "Customer Context"
        ],
        "parameters": [
          {
            "name": "rmId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/engagement/customer/{customerId}/transactions": {
      "get": {
        "operationId": "getTransactions",
        "tags": [
          "Customer Context"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/customerId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/engagement/customer/{customerId}/product-matching": {
      "post": {
        "operationId": "matchProducts",
        "tags": [
          "Customer Context"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/customerId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductMatch"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/engagement/customer/{customerId}/knowledge-map": {
      "get": {
        "operationId": "getAssembledKnowledgeMap",
        "tags": [
          "Customer Context"
        ],
        "summary": "KERT assembled knowledge map for P38",
        "description": "Calls skill-customer-previsit-report. Empty sections when KERT does not return. Does not serve the GITS knowledge-architecture snapshot.",
        "parameters": [
          {
            "$ref": "#/components/parameters/customerId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssembledKnowledgeMap"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "customerId": {
        "name": "customerId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "业务冲突",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "required": [
          "status",
          "error",
          "message",
          "path",
          "timestamp"
        ],
        "properties": {
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateCustomerRequest": {
        "type": "object",
        "required": [
          "customerId",
          "customerName"
        ],
        "properties": {
          "customerId": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "enterpriseScale": {
            "type": "string"
          },
          "customerTier": {
            "type": "string"
          },
          "rmId": {
            "type": "string"
          },
          "rmName": {
            "type": "string"
          },
          "riskLevel": {
            "type": "string"
          },
          "listedStatus": {
            "type": "string"
          }
        }
      },
      "CustomerCreatedResponse": {
        "type": "object",
        "required": [
          "customerId",
          "customerName"
        ],
        "properties": {
          "customerId": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Customer": {
        "type": "object",
        "required": [
          "customerId",
          "customerName"
        ],
        "properties": {
          "customerId": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "enterpriseScale": {
            "type": "string"
          },
          "customerTier": {
            "type": "string"
          },
          "rmId": {
            "type": "string"
          },
          "rmName": {
            "type": "string"
          },
          "riskLevel": {
            "type": "string"
          },
          "listedStatus": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CustomerOperatingView": {
        "type": "object",
        "required": [
          "customerId",
          "customerName"
        ],
        "properties": {
          "customerId": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "enterpriseScale": {
            "type": "string"
          },
          "customerTier": {
            "type": "string"
          },
          "riskLevel": {
            "type": "string"
          },
          "rmId": {
            "type": "string"
          },
          "rmName": {
            "type": "string"
          },
          "knownKycItems": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "partialKycItems": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "unknownKycItems": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "activeSignals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpportunitySignalSummary"
            }
          },
          "totalInteractions": {
            "type": "integer"
          },
          "lastInteractionTime": {
            "type": "string",
            "format": "date-time"
          },
          "activeJourneyCount": {
            "type": "integer"
          },
          "currentJourneyPhase": {
            "type": "string"
          },
          "pendingCommitments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommitmentSummary"
            }
          },
          "openReconciliationCount": {
            "type": "integer"
          },
          "riskIndicators": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "OpportunitySignalSummary": {
        "type": "object",
        "properties": {
          "signalType": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "CommitmentSummary": {
        "type": "object",
        "properties": {
          "commitmentType": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "fulfilled": {
            "type": "boolean"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "required": [
          "transactionId",
          "customerId"
        ],
        "properties": {
          "transactionId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "transactionType": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string"
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time"
          },
          "counterparty": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ProductMatch": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "matchScore": {
            "type": "number",
            "format": "double"
          },
          "matchReasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "productCategory": {
            "type": "string"
          }
        }
      },
...
      "AssemblyTraceStep": {
        "type": "object",
        "required": [
          "phase",
          "status",
          "message"
        ],
        "properties": {
          "phase": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "kiId": {
            "type": "string"
          }
        }
      },
      "SkillReportSection": {
        "type": "object",
        "properties": {
          "heading": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      },
...
      "AssembledKnowledgeMap": {
        "type": "object",
        "required": [
          "customerId",
          "skillSections",
          "assemblyTrace"
        ],
        "properties": {
          "customerId": {
            "type": "string"
          },
          "skillReportTitle": {
            "type": "string"
          },
          "skillExecutiveSummary": {
            "type": "string"
          },
          "skillSections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillReportSection"
            }
          },
          "assemblyTrace": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssemblyTraceStep"
            }
          }
        }
      },
...
    }
  }
}